[FIX] http: rewind file upload on serialization failure
Before this commit:
When uploading a file, if the transaction fails due to a serialization
failure, Odoo will retry the request. However, if a file upload is read
during the transaction, the file pointer will be at the end of the file,
and calling `.read()` again returns an empty bytes object.

After this commit:
Upon retrying the request, rewind uploads to the beginning of the file,
if the file supports it.

opw-3228200

closes odoo/odoo#119613

X-original-commit: a856c4659558ef848a3b5a0cb37d5614b16b14b2
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>
Signed-off-by: Walravens Mathieu (wama) <wama@odoo.com>